survey demo include add、update、delete、list use restful have pages

+++ src/main/java/com/roachf/survey/dao/DemoDao.java
@@ -2,6 +2,6 @@ package com.roachf.survey.dao;
 
 import com.roachf.survey.pojo.entity.Demo;
 
-public interface DemoDao extends BaseDao<Demo, Long>{
+public interface DemoDao extends BaseDao<Demo, Integer>{
 
 }
