+++ src/main/java/com/roachf/survey/service/DemoService.java
@@ -2,6 +2,6 @@ package com.roachf.survey.service;
 
 import com.roachf.survey.pojo.entity.Demo;
 
-public interface DemoService extends BaseService<Demo, Long>{
+public interface DemoService extends BaseService<Demo, Integer>{
 
 }
